Rename core classes for improved naming consistency and package layout.
