+++ byteclasses/handlers/executables/pe/dos_hdr.py
@@ -1,6 +1,6 @@
 """Legacy DOS Header Class."""
 
-from ....types.collections import FixedArray, member, structure
+from ....types.collections import ByteArray, member, structure
 from ....types.primitives.generics import Word
 from ....types.primitives.integers import Ptr32, UInt16
 
@@ -50,8 +50,8 @@ class DOSHdr:
     e_cs: UInt16  # Initial (relative) CS value
     e_lfarlc: UInt16  # File address of relocation table
     e_ovno: UInt16  # Overlay number
-    e_res: FixedArray = member(factory=lambda byte_order: FixedArray(4, UInt16))  # type: ignore
+    e_res: ByteArray = member(factory=lambda byte_order: ByteArray(4, UInt16))  # type: ignore
     e_oemid: UInt16  # OEM identifier (for e_oeminfo)
     e_oeminfo: UInt16  # OEM information; e_oemid specific
-    e_res2: FixedArray = member(factory=lambda byte_order: FixedArray(10, UInt16))  # type: ignore
+    e_res2: ByteArray = member(factory=lambda byte_order: ByteArray(10, UInt16))  # type: ignore
     e_lfanew: Ptr32  # File address of new exe header
