+++ byteclasses/types/collections/__init__.py
@@ -1,11 +1,11 @@
 """Fixed length colection types."""
 
+from .byte_array import ByteArray
+from .byteclass_collection_protocol import ByteclassCollection
 from .factory import make_fixed_collection
-from .fixed_array import FixedArray
-from .fixed_size_collection_protocol import FixedSizeCollection
 from .member import member
 from .string import String
 from .structure import structure
 from .union import union
 
-__all__ = ["FixedArray", "FixedSizeCollection", "String", "make_fixed_collection", "structure", "union", "member"]
+__all__ = ["ByteArray", "ByteclassCollection", "String", "make_fixed_collection", "structure", "union", "member"]
