+++ docs/collections/index.rst
@@ -5,7 +5,7 @@ Collections
 .. toctree::
    :maxdepth: 2
 
-   fixed_array
+   byte_array
    string
    structure
    union
