+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "byteclasses"
-version = "0.5.0"
+version = "0.6.0"
 description = "A python package designed for creating and managing blobs of binary data using dataclass like structures."
 keywords = ["struct", "structures", "binary", "blobs", "raw", "bytes", "dataclasses"]
 repository = "https://github.com/thetacom/byteclasses"
@@ -51,6 +51,7 @@ sphinx-material    = "^0.0.36"
 myst-parser        = "^2.0.0"
 nbsphinx           = "^0.9.3"
 nbqa = {extras = ["toolchain"], version = "^1.8.5"}
+poetry-plugin-export = "^1.7.1"
 
 
 [tool.black]
