+++ tests/collections/test_byte_array.py
@@ -1,25 +1,25 @@
-"""Test suite for FixedArray Byteclass."""
+"""Test suite for ByteArray Byteclass."""
 
-from byteclasses.types.collections.fixed_array import FixedArray
+from byteclasses.types.collections.byte_array import ByteArray
 from byteclasses.types.primitives.integers import UInt8, UInt32
 
 NULL_BYTE = b"\x00"
 
 
-def test_fixed_array_creation():
-    """Test FixedArray creation and properties."""
+def test_byte_array_creation():
+    """Test ByteArray creation and properties."""
     expected_count = 8
     expected_length = expected_count * len(UInt8())
-    fa = FixedArray(expected_count)
+    fa = ByteArray(expected_count)
     assert len(fa) == expected_count
     assert fa.data == NULL_BYTE * expected_length
 
 
-def test_fixed_array_creation_with_type():
-    """Test FixedArray creation and properties."""
+def test_byte_array_creation_with_type():
+    """Test ByteArray creation and properties."""
     expected_count = 8
     item_type = UInt32
     expected_length = expected_count * len(item_type())
-    fa = FixedArray(expected_count, item_type)
+    fa = ByteArray(expected_count, item_type)
     assert len(fa) == expected_length
     assert fa.data == NULL_BYTE * expected_length
