--- tests/primitives/floats/__init__.py
@@ -1,1 +1,0 @@
-"""Unit tests for byteclasses primitive float types."""
